K8S-435 Create build of the amphora image and artifact it for use
This will build an amphora imnage and upload it to the artifact
server in a directory with the RPC version and the git-sha of
the octavia repo, e.g.
images/amphora/r14.4.0/02926d8907335a4108998034bafb757418f4b5ac/

It also includes test with the new image which are used for both
post-merge and as a merge-gate.

It is implements as post-gate hook which will run periodically,
build a new image, and test it against an Octavia AIO.