RE-1550 Handle snapshot deployment (#50)
This commit updates gating/scripts/pre.sh to skip the deployment of
rpc-openstack if we're building from a snapshotted image. We also
adjust some timeouts since booting from a snapshot (which requires the
use of legacy public cloud flavours) takes a little longer than
builds that happen on standard images (which build with the performance
flavours).

Lastly, we adjust playbooks/templates/octavia.yml.aio.j2 to use
groups['shared-infra_hosts'][0] instead of ansible_hostname since when
building on a snapshot image the hostname of the build will not match
what's in the inventory.