rpc-r octavia doc and script migration
We are working on moving the octavia install docs and scripts
out of the kaas repo. Older docs and scripts were moved into
this repo using the same directory structure as kaas. The rpc-r
docs for osp 13 have been updated to include the octavia config
steps.

Working out of git issue: rcbops/rpc-prodeng#7